restricted page

+++ base/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render,redirect
 from django.contrib import messages
 from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Room,Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
@@ -11,6 +12,8 @@ from django.contrib.auth import authenticate,login,logout
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -47,7 +50,7 @@ def room(request,pk):
     
     contex={'room':room}
     return render(request,'base/room.html',contex)
-
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -58,9 +61,12 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user!=room.host:
+        return HttpResponse("why are you here?")
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -69,8 +75,11 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse("why are you here?")
     if request.method=='POST':
         room.delete()
         return redirect('home')
